feat: 로그인기능

+++ src/main/java/com/network/calendar_api/repository/EventRepository.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 List<Event> findByDateBetween(LocalDate start, LocalDate end);
-
 }
